Qrome - updated to cast value to long for compatibility in newer ESP8266 core 2.5.X

+++ printermonitor/TimeClient.cpp
@@ -148,5 +148,5 @@ long TimeClient::getCurrentEpoch() {
 }
 
 long TimeClient::getCurrentEpochWithUtcOffset() {
-  return round(getCurrentEpoch() + 3600 * myUtcOffset + 86400L) % 86400L;
+  return (long)round(getCurrentEpoch() + 3600 * myUtcOffset + 86400L) % 86400L;
 }
